Reorganized navigation files
