+++ App.js
@@ -1,26 +1,10 @@
 import "react-native-gesture-handler";
-import { StatusBar } from "expo-status-bar";
-import { StyleSheet, Text, View } from "react-native";
+import { StyleSheet } from "react-native";
 import { NavigationContainer } from "@react-navigation/native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs"; 
-import { MaterialCommunityIcons } from "@expo/vector-icons"; 
-
-import HomeScreen from "./screens/HomeScreen";
-import CreateAccountScreen from "./screens/CreateAccountScreen";
-import AddPet from "./screens/AddPet";
-import LoginScreen from "./screens/LoginScreen"; 
-import ForgotPasswordScreen from "./screens/ForgotPasswordScreen";
-import ResetPasswordScreen from "./screens/ResetPasswordScreen";
-import NewsPage from "./screens/NewsPage";
-import PetProfile from "./screens/PetProfile";
-import MenuPublicScreen from "./screens/MenuPublicScreen";
-import MenuAdminScreen from "./screens/MenuAdminScreen";
 
 import AuthNavigator from "./navigation/AuthNavigator";
-
-import colors from "./config/colors";
-
+import AppAdminNavigator from "./navigation/AppAdminNavigator";
+import AppPublicNavigator from "./navigation/AppPublicNavigator";
 
 const styles = StyleSheet.create({
   container: {
@@ -31,105 +15,10 @@ const styles = StyleSheet.create({
   },
 });
 
-const Stack = createNativeStackNavigator();
-
-const NewsPageNavigator = () => (
-  <Stack.Navigator
-    screenOptions={{
-      headerShown: false
-    }}
-  >
-    <Stack.Screen
-      name="NewsPage"
-      component={NewsPage}
-      options={{ headerShown: false }}
-    />
-  </Stack.Navigator>
-)
-
-const HomeNavigator = () => (
-  <Stack.Navigator
-    screenOptions={{
-      headerBackButtonMenuEnabled: true, 
-      headerTitle: ""
-    }}
-  >
-    <Stack.Screen
-      name="HomeStack"
-      component={HomeScreen}
-    />
-    <Stack.Screen
-      name="CreateAccount"
-      component={CreateAccountScreen}
-    />
-    <Stack.Screen
-      name="Login"
-      component={LoginScreen}
-    />
-    <Stack.Screen
-      name="ForgotPassword"
-      component={ForgotPasswordScreen}
-    />
-    <Stack.Screen
-      name="ResetPassword"
-      component={ResetPasswordScreen}
-    /> 
-    <Stack.Screen
-      name="AddPet"
-      component={AddPet}
-    />
-  </Stack.Navigator>
-)
-
-const AccountNavigator = () => (
-  <Stack.Navigator
-    screenOptions={{
-      headerBackButtonMenuEnabled: true,
-      headerTitle: ""
-    }}
-  >
-    <Stack.Screen
-      name="MenuAdmin"
-      component={MenuAdminScreen}
-    />  
-    <Stack.Screen
-      name="PetProfile"
-      component={PetProfile}
-    /> 
-  </Stack.Navigator>
-)
-
-const Tab = createBottomTabNavigator(); 
-const TabNavigator = () => (
-  <Tab.Navigator
-    screenOptions={{
-      tabBarActiveBackgroundColor: colors.lightgray, 
-      tabBarActiveTintColor: colors.black, 
-      headerShown: false
-    }}
-  >
-    <Tab.Screen 
-      name="News Feed" 
-      component={NewsPageNavigator}
-      options={{ tabBarIcon: ({ size }) => <MaterialCommunityIcons name="newspaper" size={size} />}}
-    />
-    <Tab.Screen 
-      name="Home" 
-      component={HomeNavigator} 
-      options={{ tabBarIcon: ({ size }) => <MaterialCommunityIcons name="home" size={size} />}}
-    />
-    <Tab.Screen 
-      name="Account"  
-      component={AccountNavigator}
-      options={{ tabBarIcon: ({ size }) => <MaterialCommunityIcons name="account" size={size} />}}
-    />
-  </Tab.Navigator>
-)
-
 function App() {
   return (
     <NavigationContainer>
-      <AuthNavigator />
+      <AppPublicNavigator />
     </NavigationContainer>
   );
 }
